backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
groq>=0.30.0,<1.0.0
python-dotenv==1.0.1
scikit-learn==1.5.2
numpy==2.2.0
pydantic==2.10.6
python-multipart==0.0.9
httpx==0.27.0
sqlalchemy==2.0.36
